Fixed XOR bug in crypto17, finished crypto18
In crypto17, some randomly-generated key values were producing cipher
texts that required the XORing of the same value with itself in order
to mount the padding attack. This was generating XOR results in which
leading ’00’ was dropped. In this case, the leading ’00’ is now
prepended to the result in order to proceed with the attack.